requirements.txt
pandas
openpyxl
python-docx
requests
beautifulsoup4

